refactor around singleton download and upload manager.
